+++ crates/swc_interop_nodejs/src/lib.rs
@@ -1,2 +1,2 @@
 pub mod js_hook;
-pub mod json;
+pub mod types;
